requirements.txt
numpy
pandas
scikit-learn
requests
flask
pyyaml